Move core of _standardize_field_name() to macros:
Goal: write push_header() in XS.

Problem: push_header uses _standardize_field_name() (as many other
functions do as well) and the only way to call an XS function from
another XS function is to go through the Perl stack - that's an
expensive operation.

Solution: Convert the core of _standardize_field_name() to a
purely C function. Then make the XS function call this core
function.

That's difficult since the core functionality itself requires a
few XS macros and Perl functions that work in an XS context.

Instead, this patch converts it to two macros (for readability's
sake). Another macro that calls both can be easily introduced.

The problem is that both macros require specific variables to be
provided and I was not able to introduce them in a macro like
"dMY_CXT".

This is the first step towards converting additional functions
that need to call _standardize_field_name().