Mortalize value in XS
Without the mortalization, we're leaking an SV. The Perl argument stack
does not own references to things (for efficiency). That means that our
code has the responsibility to take care of its life cycle.

Note that simply adding a SvREFCNT_dec at the end of the function would
NOT be enough. There's things that can throw an exception (here
obviously the sub call) and that would still mean a leak.

In general, I recommend making sure things are mortalized as early as
possible to minimize the latter issue.